// src/shared/types.d.ts
type DropdownMenuProps = {
	title: string;
	items: FilterData[];
};
interface FilterData {
	key: string;
	label: string;
	value: string;
	type: string;
}